Add files via upload
The files in this archive have names in the format S#P#, which stands for Sketch # Point #. These names reference the sketch number in the book, How to Cheaply Monitor and Automate Your Aquaponic/Hydroponic Garden with Arduino/ Genuino. For example, a file called S3P1 would be Sketch 3.1 in the book.

Copy or move the AquaponicsAutomation folder to your Documents/Arduino folder and it should show up in Sketchbook under the Arduino File menu. You may need to shut down and restart the IDE if it was running when you added the AquaponicsAutomation folder to the Documents/Arduino folder.